Refactor stream app.
